requirements.txt
matplotlib
numpy
scikit-learn
beautifulsoup4
lxml
python-nmap
scipy
wxpython
openai>=1.0.0
pyecharts
robotframework~=7.0
robotframework-seleniumlibrary~=6.0
playwright
pygame; python_version < "3.14"
pytest
pytest-html
pytest-rerunfailures
allure-pytest
pytest-result-log
